+++ CHANGELOG.md
@@ -13,6 +13,7 @@ It also migrates `deimos_website` into this repo.
 * deimos
     * Update Butter2 calc to use deimos_numerics delta-form SOS filter with 64-bit floats
     * Minor cleanup of construction of rev7 standard calcs
+    * Add import error on detection of irresponsible error suppression via python -O
 * deimos_numerics
     * !Add #[non_exhaustive] on `LtiError` and an error variant to support analog helpers
     * Add `DeltaSOS::set_steady_state` for initializing filters to steady state